fix Health Check Form
