+++ lib/views/screens/auth/auth_service.dart
@@ -1,6 +1,4 @@
-// import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-// import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:health_care/views/screens/auth/register/otp_popup.dart';
