+++ lib/views/screens/home/home_screens.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:convex_bottom_bar/convex_bottom_bar.dart';
 import 'package:health_care/common/app_colors.dart';
 import 'package:health_care/common/app_icons.dart';
+import 'package:health_care/views/screens/HealthCheckForm/health_check.dart';
 import 'package:health_care/views/screens/clinic/clinic_screen.dart';
 
 class HomeScreens extends StatefulWidget {
@@ -23,7 +24,7 @@ class _HomeScreensState extends State<HomeScreens> {
       case 2:
         return Center(child: Text('Bản đồ đang được phát triển'));
       case 3:
-        return Center(child: Text('Phiếu khám đang được phát triển'));
+        return HealthCheckForm();
       case 4:
         return Center(child: Text('Tài khoản đang được phát triển'));
       default:
